add method to validate ln address
If i want to use `.validate()` on my lnaddress, i need to add the `validator` crate to my project. 

Adding this method allows me to validate using the zebedee library instead of having to add an aditional crate. 

This was the error i was getting without using this method

```
error[E0599]: no method named `validate` found for struct `LnAddress` in the current scope
   --> src/player.rs:121:56
    |
121 | ...                   let valid = ln_address.validate();
    |                                              ^^^^^^^^ method not found in `LnAddress`
    |
   ::: /Users/stu/.cargo/registry/src/index.crates.io-6f17d22bba15001f/validator-0.16.1/src/traits.rs:194:8
    |
194 |     fn validate(&self) -> Result<(), ValidationErrors>;
    |        -------- the method is available for `LnAddress` here
    |
    = help: items from traits can only be used if the trait is in scope
help: the following trait is implemented but not in scope; perhaps add a `use` for it:
    |
1   + use validator::traits::Validate;
```